refact: Organize log configuration structure
